+++ src/gui.js
@@ -3395,10 +3395,11 @@ IDE_Morph.prototype.toggleRetina = function () {
 
 IDE_Morph.prototype.defaultLooks = function () {
     this.setDefaultDesign();
-    this.setDefaultTheme(this.lastDarkModeColor);
+    this.setDefaultTheme(new Color(10, 10, 10));
     this.refreshIDE();
     this.removeSetting('design');
     this.removeSetting('theme');
+    this.removeSetting('dark-mode-color');
 };
 
 IDE_Morph.prototype.flatBrightLooks = function () {
@@ -3407,6 +3408,7 @@ IDE_Morph.prototype.flatBrightLooks = function () {
     this.refreshIDE();
     this.saveSetting('design', 'flat');
     this.saveSetting('theme', 'bright');
+    this.removeSetting('dark-mode-color');
 };
 
 IDE_Morph.prototype.defaultDesign = function () {
@@ -3425,12 +3427,14 @@ IDE_Morph.prototype.defaultTheme = function () {
     this.setDefaultTheme(this.lastDarkModeColor);
     this.refreshIDE();
     this.removeSetting('theme');
+    this.removeSetting('dark-mode-color');
 };
 
 IDE_Morph.prototype.brightTheme = function () {
     this.setBrightTheme();
     this.refreshIDE();
     this.saveSetting('theme', 'bright');
+    this.removeSetting('dark-mode-color');
 };
 
 IDE_Morph.prototype.selectDarkModeColor = function() {
@@ -3466,8 +3470,10 @@ IDE_Morph.prototype.selectDarkModeColor = function() {
             null,
             () => {
                 IDE_Morph.prototype.setDefaultTheme(color);
-                this.lastDarkModeColor = color;
+                myself.lastDarkModeColor = color;
                 myself.refreshIDE();
+
+                myself.saveSetting('dark-mode-color', color.toString());
             }
         );
 
@@ -3509,8 +3515,10 @@ IDE_Morph.prototype.selectDarkModeColor = function() {
             null,
             (color) => {
                 IDE_Morph.prototype.setDefaultTheme(color);
-                this.lastDarkModeColor = color;
+                myself.lastDarkModeColor = color;
                 myself.refreshIDE();
+
+                myself.saveSetting('dark-mode-color', color.toString());
             }
         ).promptRGB(
             'Custom color',
@@ -3557,6 +3565,7 @@ IDE_Morph.prototype.applySavedSettings = function () {
 
     var design = this.getSetting('design'),
         theme = this.getSetting('theme'),
+        darkModeColor = this.getSetting('dark-mode-color'),
         zoom = this.getSetting('zoom'),
         fade = this.getSetting('fade'),
         glow = this.getSetting('glow'),
@@ -3581,7 +3590,12 @@ IDE_Morph.prototype.applySavedSettings = function () {
     if (theme === 'bright') {
         this.setBrightTheme();
     } else {
-        this.setDefaultTheme(new Color(10, 10, 10));
+        // dark mode color
+        if (darkModeColor) {
+            this.setDefaultTheme(Color.fromString(darkModeColor));
+        } else {
+            this.setDefaultTheme(new Color(10, 10, 10));
+        }
     }
 
     // blocks zoom